Move libs to services and utils to common. Webpack plugin to load tsconfig paths mappings
