+++ frontends/mit-open/package.json
@@ -22,6 +22,7 @@
     "sass-loader": "^13.3.0",
     "style-loader": "^3.3.1",
     "swc-loader": "^0.2.3",
+    "tsconfig-paths-webpack-plugin": "^4.1.0",
     "webpack": "^5.89.0",
     "webpack-bundle-analyzer": "^4.6.1",
     "webpack-bundle-tracker": "^1.4.0",
