+++ frontends/mit-open/src/api/learning-resources/resources.ts
@@ -1,5 +1,5 @@
 import type { CourseTopic, LearningResource } from "ol-search-ui"
-import axios from "../../libs/axios"
+import axios from "services/axios"
 import { useQuery } from "@tanstack/react-query"
 import type { UseQueryOptions } from "@tanstack/react-query"
 import { urls, keys } from "./urls"
