+++ frontends/mit-open/src/components/RestrictedRoute/RestrictedRoute.test.tsx
@@ -5,7 +5,7 @@ import {
   screen,
 } from "../../test-utils"
 import RestrictedRoute from "./RestrictedRoute"
-import { ForbiddenError, Permissions } from "../../util/permissions"
+import { ForbiddenError, Permissions } from "common/permissions"
 import { allowConsoleErrors } from "ol-util/test-utils"
 
 test("Renders children if permission check satisfied", () => {
