+++ frontends/mit-open/src/components/RestrictedRoute/RestrictedRoute.tsx
@@ -1,10 +1,6 @@
 import React from "react"
 import { Outlet } from "react-router"
-import {
-  ForbiddenError,
-  Permissions,
-  hasPermission,
-} from "../../util/permissions"
+import { ForbiddenError, Permissions, hasPermission } from "common/permissions"
 
 type RestrictedRouteProps = {
   children?: React.ReactNode
