+++ frontends/mit-open/src/example-request-mocks.test.ts
@@ -1,4 +1,4 @@
-import axios from "./libs/axios"
+import axios from "services/axios"
 import { setMockResponse } from "./test-utils"
 import { ControlledPromise, allowConsoleErrors } from "ol-util/test-utils"
