+++ frontends/mit-open/src/pages/ErrorPage/ErrorPage.test.tsx
@@ -7,7 +7,7 @@ import ErrorPage from "./ErrorPage"
 import { setMockResponse, mockAxiosInstance as axios } from "api/test-utils"
 import { allowConsoleErrors } from "ol-util/test-utils"
 import RestrictedRoute from "components/RestrictedRoute/RestrictedRoute"
-import { Permissions } from "../../util/permissions"
+import { Permissions } from "common/permissions"
 
 /**
  * Renders an erroring-component within a react-router ErrorBoundary using
