+++ frontends/mit-open/src/services/react-query/react-query.ts
@@ -1,5 +1,5 @@
 import { QueryClient } from "@tanstack/react-query"
-import axios from "./axios"
+import axios from "services/axios"
 
 type MaybeHasStatus = {
   response?: {
