+++ frontends/mit-open/tsconfig.json
@@ -7,7 +7,8 @@
     "paths": {
       "common/*": ["common/*"],
       "components/*": ["components/*"],
-      "page-components/*": ["page-components/*"]
+      "page-components/*": ["page-components/*"],
+      "services/*": ["services/*"]
     }
   },
   "include": ["./src/**/*.ts", "./src/**/*.tsx"],
